service_layer_handler_test_Cases

+++ application/backend/src/main/java/dao/interfaces/MeetingDaoIntf.java
@@ -1,16 +1,14 @@
 package dao.interfaces;
 
-import exceptions.EntityNotFoundException;
 import models.entities.Amenity;
 import models.entities.Meeting;
-import models.entities.Room;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
 public interface MeetingDaoIntf {
     public int bookMeeting(Meeting meeting,  int employeeId, List<Amenity>amenities);
-    public double calculateRatePerHour(Meeting meeting);
+    public int calculateRatePerHour(Meeting meeting);
     Meeting getMeetingById(int meetingId);
     List<Meeting> getAllMeetings();
